Fixed Typo
Switzerland was not written properly. Furthermore, this chart is not really representing the reality, Japan and South Korea for example do not have an unfavorable stance towards crypto at all, in fact they are both pushing adoption harder than most other countries (especially Japan). Of course regulation is increasing, but they are not going to ban crypto per se, they are trying to prevent fraud and illegal activities by unregulated exchanges, therefore the "bans" are not towards crypto but towards bad exchange practices and selling of unregistered securities (what many ERC20 tokens are).